# Conflicts:
#	modules/domain/entity/src/main/java/com/doneit/ascend/domain/entity/GroupEntity.kt
#	modules/features/featureMain/src/main/java/com/doneit/ascend/presentation/main/master_mind/list/ListViewModel.kt
#	modules/features/featureMain/src/main/res/values/strings.xml

+++ modules/storageRemote/src/main/java/com/doneit/ascend/source/storage/remote/data/response/OwnerResponse.kt
@@ -6,6 +6,6 @@ data class OwnerResponse(
     @SerializedName("id") val id: Long,
     @SerializedName("full_name") val fullName: String,
     @SerializedName("image") val image: ImageResponse,
-    @SerializedName("rating") val rating: Int,
+    @SerializedName("rating") val rating: Float,
     @SerializedName("followed") val followed: Boolean
 )
